Scaling tests for optical depths
